refactor: adapt directory structure to better mongodb integration
