+++ .gitignore
@@ -10,6 +10,7 @@ pnpm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 lerna-debug.log*
+!/src/models/logs
 
 # OS
 .DS_Store
