+++ src/common/config.ts
@@ -1,0 +1,3 @@
+export class CommonConfig {
+  static mongoConnectionURL: string = 'mongodb://127.0.0.1:27017/photos-log';
+}
