+++ src/models/logs/interfaces/log.interface.ts
