+++ src/models/logs/log.config.ts
@@ -1,0 +1,6 @@
+export class LogConfig {
+  static serviceDescribe: string = 'logsService';
+  static controllerDescribe: string = 'LogsController';
+  static modelSchemaDefinition: string = 'log';
+  static controllerDefinition: string = 'logs';
+}
