+++ src/models/logs/schemas/log.schema.ts
@@ -1,6 +1,6 @@
 import { Schema } from 'mongoose';
 
-import { MachineSchema } from './machine.schema';
+import { MachineSchema } from '../../machines/schemas/machine.schema';
 
 export const LogSchema = new Schema({
   machine: MachineSchema,
