+++ src/models/machines/dto/machine.dto.ts
