+++ src/models/machines/schemas/machine.schema.ts
